fix(Z.A.3): JIT binding-hash 64-bit (no u32 truncation)
`JitBindingEntry.name_idx` widened u32 → u64 to hold a full 64-bit
FNV-1a hash. Closes T0-T3-010 / MTT-TI-028: previous u32 truncation
(`name_hash as u32`) gave a ~4× collision rate on PLN's long freshened
variable names (`$__fr_E_BestCandidate_Helper_*`).

Sites updated:
- types/binding.rs: JitBindingEntry { name_idx: u64, ... }
- runtime/space_ops.rs:614-627: drop final `as u32` cast on hash.
- runtime/bindings.rs:45-260: remove `name_idx_u32` shim across 4
  load/store/has/exists paths; compare entry.name_idx == name_idx (u64).
- runtime/rule_dispatch.rs:148: hash_string returns u64; no cast.
- runtime/special_forms.rs:222: same.

Verification:
- New `tests/jit_binding_hash.rs` (4 tests): u64 round-trip,
  high-bit preservation (would alias under u32), 4096-name freshened
  workload with 0 collisions (vs ~4 expected under u32).
- 4149/4149 nextest pass.

+++ src/backend/bytecode/jit/runtime/rule_dispatch.rs
@@ -138,10 +138,11 @@ pub unsafe extern "C" fn jit_runtime_try_rule(ctx: *mut JitContext, rule_idx: u6
             frame.entries_count = 0;
             frame.scope_depth = ctx_ref.binding_frames_count as u32;
 
-            // Install each binding
+            // Install each binding.
+            // Z.A.3 (2026-05-12): name_idx is the full 64-bit FNV-1a hash;
+            // no truncation. hash_string returns u64 (verified at call site).
             for (name, value) in rule.bindings.iter() {
-                // We need to store the variable name index - for now store as hash
-                let name_idx = hash_string(name) as u32;
+                let name_idx = hash_string(name);
                 let jit_value = metta_to_jit(value);
 
                 let entry_ptr = frame.entries.add(frame.entries_count);
